% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microKirchhoffPlate.R
\name{ExpandLoadMKPlate}
\alias{ExpandLoadMKPlate}
\title{Expanded vector of the equivalent load}
\usage{
ExpandLoadMKPlate(tdof, elementloadMatrix, i, j, k, m)
}
\arguments{
\item{tdof}{Total degree of freedom.}

\item{elementloadMatrix}{The unexpanded vector of equivalent loads.}

\item{i}{Index of the first node.}

\item{j}{Index of the second node.}

\item{k}{Index of the third node.}

\item{m}{Index of the fourth node.}
}
\value{
Expanded vector (a column matrix) of equivalent loads.
}
\description{
This function generates the expanded vector of the equivalent load
                                            (Euler-Bernoulli beam).
}
