% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microKirchhoffPlate.R
\name{ExpandStiffnessMKPlate}
\alias{ExpandStiffnessMKPlate}
\title{Expanded stiffness matrix of a microscale Kirchhoff plate}
\usage{
ExpandStiffnessMKPlate(tdof, eMatrix, i, j, k, m)
}
\arguments{
\item{tdof}{Total degree of freedom in a
connected system of plates.}

\item{eMatrix}{Element's stiffness matrix.}

\item{i}{Index of the first node.}

\item{j}{Index of the second node.}

\item{k}{Index of the third node.}

\item{m}{Index of the fourth node.}
}
\value{
The expanded stiffness matrix
                     a microscale Kirchhoff plate element.
}
\description{
Accepts the element's stiffness matrix and returns the expanded form
of the matrix to conform with the global degree of freedom.
}
