% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaryConditions.R, R/microTimoshenkoBeam.R
\name{ExtractFreeRows}
\alias{ExtractFreeRows}
\title{Rows of unrestrained nodes.}
\usage{
ExtractFreeRows(total_DOF, restrainednodes)

ExtractFreeRows(total_DOF, restrainednodes)
}
\arguments{
\item{total_DOF}{A single number indicating the
total degree of freedom in a
connected system of beams.}

\item{restrainednodes}{A vector of restrained nodes.}

\item{total_DOF}{A single number indicating the
total degree of freedom in a
connected system of beams.}

\item{restrainednodes}{A vector of restrained nodes.}
}
\value{
A vector of rows of known loads.
                          (or a vector of rows of unrestrained nodes).

A vector of rows of known loads.
                          (or a vector of rows of unrestrained nodes).
}
\description{
Extracts the rows of unrestrained nodes.
This is required for reducing the global stiffness matrix.

Extracts the rows of unrestrained nodes.
This is required for reducing the global stiffness matrix.
}
