% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massStabilityMatricesMTB.R
\name{FindCriticalLoad}
\alias{FindCriticalLoad}
\title{Buckling loads (Timoshenko beam element)}
\usage{
FindCriticalLoad(reducedM, reducedK)
}
\arguments{
\item{reducedK}{Reduced mass matrix obtained by applying
boundary conditions on the global stiffness matrix}

\item{reducedS}{Reduced stability matrix obtained by applying
boundary conditions on the global stability matrix}
}
\value{
Critical buckling load
}
\description{
This function computes the buckling loads, but returns only the critical buckling load.
}
