% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microTimoshenkoBeam.R
\name{FindForcesMomentsMTB}
\alias{FindForcesMomentsMTB}
\title{Local elements forces and moments of a microscale Timoshenko beam.}
\usage{
FindForcesMomentsMTB(youngmod, shearmod, momentinertia, area, shearfactor,
  poissonratio, length, lengthscale, global_dof, i, j)
}
\arguments{
\item{youngmod}{Young's modulus.}

\item{shearmod}{Shear modulus.}

\item{momentinertia}{Moment of inertia (e.g. bh^3/12 for rectangle).}

\item{area}{Cross-section area.}

\item{shearfactor}{Shear correction factor (constant k).}

\item{poissonratio}{Poisson's ratio.}

\item{length}{Element's length.}

\item{lengthscale}{Material length scale parameter}

\item{global_dof}{Global degrees of freedom
(combination of the dof found
with FindNodalDOF() and the boundary conditions).}

\item{i}{Index of the first node.}

\item{j}{Index of the second node.}
}
\value{
Local forces and moments (MTB).
}
\description{
Solves for the nodal forces and moments of an element.
}
