% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massStabilityMatricesMTB.R
\name{FindFrequenciesRad}
\alias{FindFrequenciesRad}
\title{Natural frequencies (timoshenko beam element in radians)}
\usage{
FindFrequenciesRad(reducedM, reducedK)
}
\arguments{
\item{reducedM}{Reduced mass matrix obtained by applying
boundary condition on the global mass matrix}

\item{reducedK}{Reduced mass matrix obtained by applying
boundary condition on the global stiffness matrix}
}
\value{
Natural frequencies in radian/s
}
\description{
This function computes the natural frequencies, in radians per seconds.
}
