% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microTimoshenkoBeam.R
\name{FindNodalDOFs}
\alias{FindNodalDOFs}
\title{Nodal degrees of freedom}
\usage{
FindNodalDOFs(reducedk, reducedforce)
}
\arguments{
\item{reducedk}{Reduced stiffness matrix
obtained with FormReducedStiffnes().}

\item{reducedforce}{Reduced force vector obtained with
FormReducedForce().}
}
\value{
A vector of all unknown nodal
                        degrees of freedom.
}
\description{
Determines the unknown nodal degrees of freedom.
}
