% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaryConditions.R, R/microTimoshenkoBeam.R
\name{FixNodes}
\alias{FixNodes}
\title{Fixed or clamped boundary condition(s)}
\usage{
FixNodes(nodes)

FixNodes(nodes)
}
\arguments{
\item{nodes}{A vector of node(s) for which the
displacements are to be suppressed.}

\item{nodes}{A vector of node(s) for which the
displacements are to be suppressed.}
}
\value{
A vector of rows (fixed nodes) to be eliminated
                    in the global matrix equation.

A vector of rows (fixed nodes) to be eliminated
                    in the global matrix equation.
}
\description{
Retrieves the translational and rotational degree of freedom
                                 of a node or a set of node.

Retrieves the translational and rotational degrees of freedom
                                 of a node or a set of node.
}
