% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaryConditions2D.R
\name{FixNodes2D}
\alias{FixNodes2D}
\title{Fixed or clamped boundary condition(s)}
\usage{
FixNodes2D(nodes)
}
\arguments{
\item{nodes}{A vector of node(s) for which the the
translational and rotational displacements are to be suppressed.}
}
\value{
A vector of rows (fixed nodes) to be eliminated
                    in the global matrix equation.
}
\description{
Retrieves the translational and rotational degrees of freedom
                                 of a node or a set of node.
}
