% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleMatrixMKP.R
\name{FormGlobalKM2D}
\alias{FormGlobalKM2D}
\title{Form global matrix from element matrix (stiffness, or mass) for a plate element}
\usage{
FormGlobalKM2D(enumber, listnodes, youngmod, poissonratio, edge_a, edge_b,
  thickness, lengthscale, rho, case)
}
\arguments{
\item{enumber}{Number of discretized plate elements.}

\item{listnodes}{A list of nodal indices for the elements}

\item{youngmod}{Young's modulus.}

\item{poissonratio}{Poisson's ratio.}

\item{edge_a}{Length along x-axis}

\item{edge_b}{Length along y-axis}

\item{thickness}{Plate thinckness}

\item{lengthscale}{Material length scale parameter}

\item{rho}{Mass density (kg/cubic meter)}

\item{case}{Used to specify if stiffness matrix or mass matrix
or stability matrix. Note that case = 1 (stiffness),
or 2 (mass), or 3(stability))}
}
\value{
Depending on what is being evaluated, it returns
                        the global stiffness matrix (case 1) or mass matrix (case 2)
}
\description{
Form global matrix from element matrix (stiffness, or mass) for a plate element
}
