% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleMatrixMTB.R
\name{FormGlobalKMS}
\alias{FormGlobalKMS}
\title{Form global matrix from element matrix (stiffness, mass and stability)}
\usage{
FormGlobalKMS(enumber, youngmod, shearmod, momentinertia, area,
  shearfactor, poissonratio, totallength, lengthscale, rho, case)
}
\arguments{
\item{enumber}{Number of discretized elements.}

\item{youngmod}{Young's modulus.}

\item{shearmod}{Shear modulus.}

\item{momentinertia}{Moment of inertia (e.g. bh^3/12 for a rectangle).}

\item{area}{Cross-sectional area}

\item{shearfactor}{Shear correction factor (constant k).}

\item{poissonratio}{Poisson's ratio.}

\item{totallength}{Total length of the undiscretized structure}

\item{lengthscale}{Material length scale parameter}

\item{rho}{Mass density (kg/cubic meter)}

\item{case}{Used to specify if stiffness matrix or mass matrix
or stability matrix. Note that case = 1 (stiffness),
or 2 (mass), or 3(stability))}
}
\value{
Depending on what is being evaluated, it returns
                        the global stiffness, or mass or stability matrix
}
\description{
Form global matrix from element matrix (stiffness, mass and stability)
}
