% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleMatrixMKP.R
\name{FormGlobalLoad2D}
\alias{FormGlobalLoad2D}
\title{Global load vector for a plate with surface UDL}
\usage{
FormGlobalLoad2D(enumber, listnodes, loadq, edgea, edgeb)
}
\arguments{
\item{enumber}{Number of discretized plate elements.}

\item{listnodes}{A list of nodal indices for the elements}

\item{edge_a}{Length along x-axis}

\item{edge_b}{Length along y-axis}
}
\value{
Returns the global vector of equivalent load
}
\description{
Form global load vector (a column matrix) for a plate with surface UDL
}
