% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microKirchhoffPlate.R
\name{FormLoadMKPlate}
\alias{FormLoadMKPlate}
\title{Equivalent load vector (non-conforming rectangular Kirchhoff plate element)}
\usage{
FormLoadMKPlate(pressuremag, edge_a, edge_b)
}
\arguments{
\item{pressuremag}{Pressure magnitude, e.g. q in N/m^2.}

\item{edge_a}{Length along x-axis}

\item{edge_b}{Length along y-axis}
}
\value{
A column matrix of the equivalent nodal loads.
}
\description{
Generates the column matrix of equivalent loads for a
          non-conforming rectangular Kirchhoff plate.
}
