% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microKirchhoffPlate.R
\name{FormMassMKplate}
\alias{FormMassMKplate}
\title{Mass matrix of a modified couple stress Kirchhoff plate element.}
\usage{
FormMassMKplate(youngmod, edge_a, edge_b, thickness, rho)
}
\arguments{
\item{youngmod}{Young's modulus}

\item{edge_a}{Length along x-axis}

\item{edge_b}{Length along y-axis}

\item{thickness}{Plate thinckness}

\item{rho}{Mass density of the plate}
}
\value{
Mass matrix (combination of the
                              classical stiffness matrix of a Timoshenko beam with
                              the effect arising from the modified
                              couple stress theory).
}
\description{
Forms the mass matrix of a microscale Kirchhoff plate element (MKP)
}
