% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massStabilityMatricesMTB.R
\name{FormMassMTB}
\alias{FormMassMTB}
\title{Mass matrix of a modified couple stress Timoshenko beam.}
\usage{
FormMassMTB(youngmod, shearmod, momentinertia, area, shearfactor,
  poissonratio, length, rho)
}
\arguments{
\item{youngmod}{Young's modulus.}

\item{shearmod}{Shear modulus.}

\item{momentinertia}{Moment of inertia (e.g. bh^3/12 for a rectangle).}

\item{area}{Cross-sectional area}

\item{shearfactor}{Shear correction factor (constant k).}

\item{poissonratio}{Poisson's ratio.}

\item{length}{Element's length.}

\item{rho}{Mass density (kg/cubic meter)}
}
\value{
Mass matrix involving both translational and
                           rotary mass matrices.
}
\description{
Forms the mass matrix of a microscale Timoshenko beam (MTB)
}
