% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microKirchhoffPlate.R
\name{FormReducedLoad2D}
\alias{FormReducedLoad2D}
\title{Reduced load vector.}
\usage{
FormReducedLoad2D(bigColumnMatrix, unrestrainednodes)
}
\arguments{
\item{bigColumnMatrix}{A vector of applied nodal loads (force,moments etc)
obtained from using FormLoadMKPlate() for a single element
or the assembled load matrix}

\item{unrestrainednodes}{A set of freenodes retrieved after applying the boundary condition.}
}
\value{
A column matrix of the reduced load vector
}
\description{
Forms the reduced load matrix after applying the boundary condition(s).
}
