% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microTimoshenkoBeam.R
\name{FormReducedMatrix}
\alias{FormReducedMatrix}
\title{Reduced stiffness or mass matrix.}
\usage{
FormReducedMatrix(globalmat, vec_freenodes)
}
\arguments{
\item{globalmat}{Global mass matrix (assembled mass matrices)}

\item{vec_freenodes}{The set of nodes of freed nodes obtained by excluding the rows
of the fixed/hinged nodes in the global dof.}
}
\value{
Reduced stiffness or reduced mass matrix
}
\description{
Given the global stiffness or mass matrix, and
the vector of free nodes, this function returns the
reduced stiffness or reduced mass matrix.
}
