% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microTimoshenkoBeam.R
\name{FormReducedStiffness}
\alias{FormReducedStiffness}
\title{Reduced stiffness matrix}
\usage{
FormReducedStiffness(globalK, knownforcenodes)
}
\arguments{
\item{globalK}{Global stiffness matrix
(assembled matrices of all elements)}

\item{knownforcenodes}{The set of nodes for the applied loads
obtained by excluding the rows
of the fixed/hinged nodes in the global dof.}
}
\value{
Reduced stiffness matrix
}
\description{
Reduced stiffness matrix
}
