% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massStabilityMatricesMTB.R
\name{FormStabilityMTB}
\alias{FormStabilityMTB}
\title{Stability matrix of a modified couple stress Timoshenko beam.}
\usage{
FormStabilityMTB(youngmod, shearmod, momentinertia, area, shearfactor,
  poissonratio, length)
}
\arguments{
\item{youngmod}{Young's modulus.}

\item{shearmod}{Shear modulus.}

\item{momentinertia}{Moment of inertia (e.g. bh^3/12 for a rectangle).}

\item{area}{Cross-sectional area}

\item{shearfactor}{Shear correction factor (constant k).}

\item{poissonratio}{Poisson's ratio.}

\item{length}{Element's length.}
}
\value{
Stability matrix for a microscale Timoshenko beam
}
\description{
Forms the stability matrix of a microscale Timoshenko beam (MTB)
}
