% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microKirchhoffPlate.R
\name{FormStiffnessMKplate}
\alias{FormStiffnessMKplate}
\title{Stiffness matrix of a modified couple stress Kirchhoff plate element.}
\usage{
FormStiffnessMKplate(youngmod, poissonratio, edge_a, edge_b, thickness,
  lengthscale)
}
\arguments{
\item{youngmod}{Young's modulus}

\item{poissonratio}{Poisson's ratio}

\item{edge_a}{Length along x-axis}

\item{edge_b}{Length along y-axis}

\item{thickness}{Plate thinckness}

\item{lengthscale}{Material length scale parameter}
}
\value{
Stiffness matrix (combination of the
                               classical stiffness matrix of a Timoshenko beam with
                               the effect arising from the modified
                               couple stress theory).
}
\description{
Forms the stiffness matrix of a microscale Kirchhoff plate element (MKP)
}
