% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microTimoshenkoBeam.R
\name{FormStiffnessMTB}
\alias{FormStiffnessMTB}
\title{Stiffness matrix of a modified couple stress Timoshenko beam.}
\usage{
FormStiffnessMTB(youngmod, shearmod, momentinertia, area, shearfactor,
  poissonratio, length, lengthscale)
}
\arguments{
\item{youngmod}{Young's modulus.}

\item{shearmod}{Shear modulus.}

\item{momentinertia}{Moment of inertia (e.g. bh^3/12 for a rectangle).}

\item{area}{Cross-section area.}

\item{shearfactor}{Shear correction factor (constant k).}

\item{poissonratio}{Poisson's ratio.}

\item{length}{Element's length.}

\item{lengthscale}{Material length scale parameter}
}
\value{
Stiffness matrix (combination of the
                           classical stiffness matrix of a Timoshenko beam with
                           the effect arising from the modified
                            couple stress theory).
}
\description{
Forms the stiffness matrix of a microscale Timoshenko beam (MTB)
}
