% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaryConditions.R, R/microTimoshenkoBeam.R
\name{HingeNodes}
\alias{HingeNodes}
\title{Hinged or simply supported boundary condition(s)}
\usage{
HingeNodes(nodes)

HingeNodes(nodes)
}
\arguments{
\item{nodes}{A vector of node(s) for which the
translational displacement is to be suppressed.}

\item{nodes}{A vector of node(s) for which the
translational displacement is to be suppressed.}
}
\value{
A vector of rows (hinged nodes) to be eliminated
                    in the global matrix equation.

A vector of rows (hinged nodes) to be eliminated
                    in the global matrix equation.
}
\description{
Retrieves the translational displacement/dof of a node or a set of node.

Retrieves the translational displacement/dof of a node or a set of node.
}
