% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaryConditions2D.R
\name{HingeNodes2Dx}
\alias{HingeNodes2Dx}
\title{Simply supported boundary condition(s) along x}
\usage{
HingeNodes2Dx(nodes)
}
\arguments{
\item{nodes}{A vector of node(s) for which the
translational displacement is to be suppressed.}
}
\value{
A vector of rows (hinged nodes) to be eliminated
                    in the global matrix equation.
}
\description{
Retrieves the displacement and theta_y of a node or a set of node along x.
}
