% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaryConditions2D.R
\name{HingeNodes2Dy}
\alias{HingeNodes2Dy}
\title{Simply supported boundary condition(s) along y}
\usage{
HingeNodes2Dy(nodes)
}
\arguments{
\item{nodes}{A vector of node(s) for which the
translational displacement is to be suppressed.}
}
\value{
A vector of rows (hinged nodes) to be eliminated
                    in the global matrix equation.
}
\description{
Retrieves the displacement and theta_x of a node or a set of node along y.
}
