% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaryConditions2D.R
\name{Symmetry2Dx}
\alias{Symmetry2Dx}
\title{Symmetric boundary condition(s) along x}
\usage{
Symmetry2Dx(nodes)
}
\arguments{
\item{nodes}{A vector of nodes for which the symmetric boundary condition(s) is to be applied}
}
\value{
A vector of rows (hinged nodes) to be eliminated
                                       in the global matrix equation.
}
\description{
Retrieves theta_y of a node or a set of node along x.
}
